+++ driver/xf86-video-intel/src/sna/fb/fbseg.c
@@ -25,7 +25,6 @@
 
 #include "fb.h"
 #include "fbclip.h"
-#include <mi.h>
 #include <miline.h>
 #include <scrnintstr.h>
 
@@ -353,7 +352,8 @@ fbSelectBres(DrawablePtr drawable, GCPtr gc)
 	FbBres *bres;
 
 	DBG(("%s: line=%d, fill=%d, and=%lx, bgand=%lx\n",
-	     __FUNCTION__, gc->lineStyle, gc->fillStyle, pgc->and, pgc->bgand));
+	     __FUNCTION__, gc->lineStyle, gc->fillStyle,
+	     (long)pgc->and, (long)pgc->bgand));
 	assert(gc->lineWidth == 0);
 
 	if (gc->lineStyle == LineSolid) {
